+++ src/main/java/org/plumelib/util/SystemPlume.java
@@ -5,8 +5,9 @@ package org.plumelib.util;
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.ArrayDeque;
+import java.util.Deque;
+import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.qual.Pure;
 
 /** Utility methods relating to the JVM runtime system: sleep and garbage collection. */
@@ -135,86 +136,6 @@ public final class SystemPlume {
     return result;
   }
 
-  // This should probably be a deque, so that it can be pruned.
-  // A problem is that a deque cannot be iterated through; it does not implement `get()`.
-  /**
-   * A list of pairs of (timestamp, cumulative collection time). The timestamp is an epoch second,
-   * and the collection time is in milliseconds. New items are added to the end.
-   *
-   * <p>The list is not currently pruned.
-   */
-  private static List<Pair<Long, Long>> gcHistory = new ArrayList<>();
-
-  /**
-   * Returns the fraction of time spent garbage collecting, in the past minute. This is generally a
-   * value between 0 and 1. This method might return a value greater than 1 if multiple threads are
-   * spending all their time collecting. Returns 0 if {@code gcPercentage} was not first called more
-   * than 1 minute ago.
-   *
-   * <p>A typical use is to put the following in an outer loop that takes a significant amount of
-   * time (more than a second) to execute:
-   *
-   * <pre>{@code
-   * if (GC.gcPercentage() > .25) {
-   *   String message = String.format(
-   *     "Garbage collection consumed over 25% of CPU during the past minute."
-   *     + " Perhaps increase max heap size (max memory = %d, total memory = %d, free memory = %d).",
-   *     Runtime.getRuntime().maxMemory(),
-   *     Runtime.getRuntime().totalMemory(),
-   *     Runtime.getRuntime().freeMemory());
-   *   System.err.println(message);
-   * }
-   * }</pre>
-   *
-   * @return the percentage of time spent garbage collecting, in the past minute
-   */
-  public static double gcPercentage() {
-    return gcPercentage(60);
-  }
-
-  /**
-   * Returns the fraction of time spent garbage collecting, in the past {@code seconds} seconds.
-   * This is generally a value between 0 and 1. This method might return a value greater than 1 if
-   * multiple threads are spending all their time collecting. Returns 0 if {@code gcPercentage} was
-   * not first called more than {@code seconds} seconds ago.
-   *
-   * <p>A typical use is to put the following in an outer loop that takes a significant amount of
-   * time (more than a second) to execute:
-   *
-   * <pre>{@code
-   * if (GC.gcPercentage(10) > .25) {
-   *   String message = String.format(
-   *     "Garbage collection consumed over 25% of CPU during the past 10 seconds."
-   *     + " Perhaps increase max heap size (max memory = %d, total memory = %d, free memory = %d).",
-   *     Runtime.getRuntime().maxMemory(),
-   *     Runtime.getRuntime().totalMemory(),
-   *     Runtime.getRuntime().freeMemory());
-   *   System.err.println(message);
-   * }
-   * }</pre>
-   *
-   * @param seconds the size of the time window, in seconds
-   * @return the percentage of time spent garbage collecting, in the past {@code seconds} seconds
-   */
-  public static double gcPercentage(int seconds) {
-    long now = Instant.now().getEpochSecond(); // in seconds
-    long collectionTime = getCollectionTime(); // in milliseconds
-    gcHistory.add(Pair.of(now, collectionTime));
-
-    for (int i = gcHistory.size() - 1; i >= 0; i--) {
-      Pair<Long, Long> p = gcHistory.get(i);
-      long historyTimestamp = p.a; // in seconds
-      long elapsed = now - historyTimestamp; // in seconds
-      if (elapsed >= seconds) {
-        long historyCollectionTime = p.b; // in milliseconds
-        double elapsedCollectionTime =
-            (collectionTime - historyCollectionTime) / 1000.0; // in seconds
-        return elapsedCollectionTime / elapsed;
-      }
-    }
-    return 0;
-  }
-
   /**
    * Returns the cumulative garbage collection time in milliseconds, across all threads.
    *
@@ -229,5 +150,134 @@ public final class SystemPlume {
       }
     }
     return result;
+  }
+
+  /** A triple of (timestamp, collection time, subsequent timestamp). */
+  private static class GcHistoryItem {
+    /** When the collection happened. An epoch second. */
+    long timestamp;
+    /** The cumulative collection time in milliseconds. */
+    long collectionTime;
+    /**
+     * When the subsequent collection happened. It is 0 until after the subsequent collection
+     * occurs. The purpose of this field is to avoid the need for a {@code peek2()} method on deque.
+     */
+    long subsequentTimestamp = 0;
+
+    /**
+     * Creates a new GcHistoryItem.
+     *
+     * @param timestamp when the collection happened; an epoch second
+     * @param collectionTime the collection time in milliseconds
+     */
+    GcHistoryItem(long timestamp, long collectionTime) {
+      this.timestamp = timestamp;
+      this.collectionTime = collectionTime;
+    }
+  }
+
+  /** The history of recent garbage collection runs. The queue is never empty. */
+  private static Deque<GcHistoryItem> gcHistory;
+
+  static {
+    gcHistory = new ArrayDeque<>();
+    // Add a dummy element so the queue is never empty.
+    gcHistory.add(new GcHistoryItem(Instant.now().getEpochSecond(), getCollectionTime()));
+  }
+
+  /**
+   * Calls `gcPercentage(60)`.
+   *
+   * @return the percentage of time spent garbage collecting, in the past minute
+   * @see #gcPercentage()
+   */
+  public static double gcPercentage() {
+    return gcPercentage(60);
+  }
+
+  /**
+   * Returns the fraction of time spent garbage collecting, in the past {@code seconds} seconds.
+   * This is generally a value between 0 and 1. This method might return a value greater than 1 if
+   * multiple threads are spending all their time collecting. Returns 0 if {@code gcPercentage} was
+   * not first called more than {@code seconds} seconds ago.
+   *
+   * <p>This method also discards all GC history older than {@code seconds} seconds.
+   *
+   * <p>Instead of calling this method directly, a client program might call {@link
+   * #gcUsageMessage}.
+   *
+   * @param seconds the size of the time window, in seconds
+   * @return the percentage of time spent garbage collecting, in the past {@code seconds} seconds
+   */
+  public static double gcPercentage(int seconds) {
+    GcHistoryItem newest = gcHistory.getLast();
+    long now = Instant.now().getEpochSecond(); // in seconds
+    long collectionTime; // in milliseconds
+    if (now == newest.timestamp) {
+      // For efficiency, don't add another entry with the same timestamp as the newest one, even
+      // though the collectionTime field would differ slightly.
+      collectionTime = newest.collectionTime;
+    } else {
+      newest.subsequentTimestamp = now;
+      collectionTime = getCollectionTime();
+      newest = new GcHistoryItem(now, collectionTime);
+      gcHistory.add(newest);
+    }
+
+    GcHistoryItem oldest = gcHistory.getFirst();
+    while (oldest.subsequentTimestamp != 0 && now - oldest.subsequentTimestamp > seconds) {
+      // The second-oldest history item can be used, so don't use the oldest one.
+      gcHistory.removeFirst();
+      oldest = gcHistory.getFirst();
+    }
+    // At this point, the second-oldest history item is too recent to use (or it does not exist).
+
+    long elapsed = now - oldest.timestamp; // in seconds
+    if (elapsed < seconds) {
+      // The oldest history item is too recent to use.
+      return 0;
+    }
+
+    double elapsedCollectionTime = (collectionTime - oldest.collectionTime) / 1000.0; // in seconds
+    return elapsedCollectionTime / elapsed;
+  }
+
+  /**
+   * If the fraction of time spent garbage collecting in the past {@code seconds} seconds is less
+   * than {@code cpuThreshold}, returns null. Otherwise, returns a string indicating garbage
+   * collection CPU usage and memory statistics. The string is multiple lines long, but does not end
+   * with a line separator.
+   *
+   * <p>A typical use is to put the following in an outer loop that takes a significant amount of
+   * time (more than a second) to execute:
+   *
+   * <pre>{@code
+   * String message = gcUsageMessage(.25, 60);
+   * if (message != null) {
+   *   System.err.println(message);
+   * }
+   * }</pre>
+   *
+   * @param cpuThreshold the maximum fraction of CPU that should be spent garbage collecting; a
+   *     number between 0 and 1
+   * @param seconds the time window in which to compute the garbage collection CPU usage
+   * @return a GC usage message string, or null
+   */
+  public static @Nullable String gcUsageMessage(double cpuThreshold, int seconds) {
+    double gcPercentage = SystemPlume.gcPercentage(seconds);
+    if (gcPercentage < cpuThreshold) {
+      return null;
+    } else {
+      return String.join(
+          System.lineSeparator(),
+          "Garbage collection consumed "
+              + Math.round(gcPercentage * 100)
+              + "% of CPU during the past "
+              + seconds
+              + " seconds.",
+          "  max memory = " + Runtime.getRuntime().maxMemory(),
+          "total memory = " + Runtime.getRuntime().totalMemory(),
+          " free memory = " + Runtime.getRuntime().freeMemory());
+    }
   }
 }