+++ src/main/java/org/plumelib/util/WeakHasherMap.java
@@ -23,7 +23,9 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Set;
+import org.checkerframework.checker.lock.qual.GuardSatisfied;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.checkerframework.checker.signedness.qual.UnknownSignedness;
 import org.checkerframework.dataflow.qual.Pure;
 import org.checkerframework.dataflow.qual.SideEffectFree;
 
@@ -104,13 +106,27 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
   done.  Fortunately these are small, short-lived objects, so the added
   allocation overhead is tolerable. */
 
+  /** hasher. */
   private Hasher hasher = null;
 
+  /**
+   * keyEquals.
+   *
+   * @param k1 k1
+   * @param k2 k2
+   * @return return
+   */
   @Pure
   private boolean keyEquals(Object k1, Object k2) {
     return (hasher == null ? k1.equals(k2) : hasher.equals(k1, k2));
   }
 
+  /**
+   * keyHashCode.
+   *
+   * @param k1 k1
+   * @return return
+   */
   @Pure
   private int keyHashCode(Object k1) {
     return (hasher == null ? k1.hashCode() : hasher.hashCode(k1));
@@ -122,36 +138,75 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
   // were static in the original version of this code.
   // This finesses that.
 
+  /**
+   * WeakKeyCreate.
+   *
+   * @param k k
+   * @return return
+   */
   private @Nullable WeakKey WeakKeyCreate(K k) {
     if (k == null) return null;
     else return new WeakKey(k);
   }
 
+  /**
+   * WeakKeyCreate.
+   *
+   * @param k k
+   * @param q q
+   * @return return
+   */
   private @Nullable WeakKey WeakKeyCreate(K k, ReferenceQueue<? super K> q) {
     if (k == null) return null;
     else return new WeakKey(k, q);
   }
 
+  /** WeakKey. */
   // Cannot be a static class: uses keyHashCode() and keyEquals()
   private final class WeakKey extends WeakReference<K> {
+    /** hash. */
     private int hash; /* Hashcode of key, stored here since the key
 				   may be tossed by the GC */
 
+    /**
+     * WeakKey.
+     *
+     * @param k k
+     */
     private WeakKey(K k) {
       super(k);
       hash = keyHashCode(k);
     }
 
+    /**
+     * create.
+     *
+     * @param k k
+     * @return return
+     */
     private @Nullable WeakKey create(K k) {
       if (k == null) return null;
       else return new WeakKey(k);
     }
 
+    /**
+     * WeakKey.
+     *
+     * @param k k
+     * @param q q
+     */
     private WeakKey(K k, ReferenceQueue<? super K> q) {
       super(k, q);
       hash = keyHashCode(k);
     }
 
+    /**
+     * create.
+     *
+     * @param k k
+     * @param q q
+     * @return return
+     */
     private @Nullable WeakKey create(K k, ReferenceQueue<? super K> q) {
       if (k == null) return null;
       else return new WeakKey(k, q);
@@ -183,10 +238,10 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
     }
   }
 
-  /* Hash table mapping WeakKeys to values */
+  /** Hash table mapping WeakKeys to values */
   private Map<WeakKey, V> hash;
 
-  /* Reference queue for cleared WeakKeys */
+  /** Reference queue for cleared WeakKeys */
   private ReferenceQueue<? super K> queue = new ReferenceQueue<K>();
 
   /* Remove all invalidated entries from the map, that is, remove all entries
@@ -194,6 +249,7 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
   each public mutator in this class.  We don't invoke this method in
   public accessors because that can lead to surprising
   ConcurrentModificationExceptions. */
+  /** processQueue. */
   @SuppressWarnings("unchecked")
   private void processQueue() {
     WeakKey wk;
@@ -291,7 +347,8 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
    */
   @Pure
   @Override
-  public @Nullable V get(Object key) { // type of argument is Object, not K
+  public @Nullable V get(
+      @UnknownSignedness @GuardSatisfied Object key) { // type of argument is Object, not K
     @SuppressWarnings("unchecked")
     K kkey = (K) key;
     return hash.get(WeakKeyCreate(kkey));
@@ -337,15 +394,23 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
 
   /* -- Views -- */
 
-  /* Internal class for entries */
+  /** Internal class for entries */
   // This can't be static, again because of dependence on hasher.
   @SuppressWarnings("TypeParameterShadowing")
   private final class Entry<K, V> implements Map.Entry<K, V> {
+    /** ent. */
     private Map.Entry<WeakKey, V> ent;
+    /** key. */
     private K key; /* Strong reference to key, so that the GC
 				   will leave it alone as long as this Entry
 				   exists */
 
+    /**
+     * Entry.
+     *
+     * @param ent ent
+     * @param key key
+     */
     Entry(Map.Entry<WeakKey, V> ent, K key) {
       this.ent = ent;
       this.key = key;
@@ -368,16 +433,36 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
       return ent.setValue(value);
     }
 
+    /**
+     * keyvalEquals.
+     *
+     * @param o1 o1
+     * @param o2 o2
+     * @return return
+     */
     @Pure
     private boolean keyvalEquals(K o1, K o2) {
       return (o1 == null) ? (o2 == null) : keyEquals(o1, o2);
     }
 
+    /**
+     * valEquals.
+     *
+     * @param o1 o1
+     * @param o2 o2
+     * @return return
+     */
     @Pure
     private boolean valEquals(V o1, V o2) {
       return (o1 == null) ? (o2 == null) : o1.equals(o2);
     }
 
+    /**
+     * equals.
+     *
+     * @param e e
+     * @return return
+     */
     @Pure
     @SuppressWarnings("NonOverridingEquals")
     public boolean equals(Map.Entry<K, V> e /* Object o*/) {
@@ -395,8 +480,9 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
     }
   }
 
-  /* Internal class for entry sets */
+  /** Internal class for entry sets */
   private final class EntrySet extends AbstractSet<Map.Entry<K, V>> {
+    /** hashEntrySet. */
     Set<Map.Entry<WeakKey, V>> hashEntrySet = hash.entrySet();
 
     @Override
@@ -482,6 +568,7 @@ public final class WeakHasherMap<K, V> extends AbstractMap<K, V> implements Map<
     }
   }
 
+  /** entrySet. */
   private @Nullable Set<Map.Entry<K, V>> entrySet = null;
 
   /** Returns a <code>Set</code> view of the mappings in this map. */