+++ src/test/java/org/plumelib/util/StringsPlumeTest.java
@@ -98,6 +98,7 @@ public final class StringsPlumeTest {
     assertEquals(s, StringsPlume.unescapeJava(escaped));
   }
 
+  @SuppressWarnings("UnicodeEscape")
   @Test
   public void test_escapeJava() {
 
@@ -145,7 +146,6 @@ public final class StringsPlumeTest {
     assertEquals("(1", StringsPlume.unescapeJava("\501"));
     assertEquals("?7", StringsPlume.unescapeJava("\0777")); // '?' = \077
     assertEquals("?7", StringsPlume.unescapeJava("\777")); // '?' = \077
-    assertEquals(" M ", StringsPlume.unescapeJava(" \uuu004D "));
 
     // public static String escapeNonASCII(String orig)